rat_frontend/requirements_rat_frontend.txt
alembic==1.18.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
crontab==1.0.5
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Principal==0.4.0
Flask-Security==5.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.4.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
libpass==1.9.3
Mako==1.3.11
Markdown==3.10.2
MarkupSafe==3.0.3
numpy==2.4.4
pandas==3.0.2
psycopg2-binary==2.9.11
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
six==1.17.0
SQLAlchemy==2.0.49
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
Werkzeug==3.1.8
WTForms==3.2.1
xlsxwriter==3.2.9
